import and test fixes
