+++ python/semantic_kernel/connectors/memory/qdrant/qdrant_collection.py
@@ -16,7 +16,8 @@ from qdrant_client.models import PointStruct, VectorParams
 
 from semantic_kernel.connectors.memory.qdrant.const import DISTANCE_FUNCTION_MAP, TYPE_MAPPER_VECTOR
 from semantic_kernel.connectors.memory.qdrant.utils import AsyncQdrantClientWrapper
-from semantic_kernel.data import VectorStoreRecordCollection, VectorStoreRecordDefinition, VectorStoreRecordVectorField
+from semantic_kernel.data.record_definition import VectorStoreRecordDefinition, VectorStoreRecordVectorField
+from semantic_kernel.data.vector_storage import VectorStoreRecordCollection
 from semantic_kernel.exceptions import (
     MemoryConnectorInitializationError,
     VectorStoreModelValidationError,
