+++ python/semantic_kernel/connectors/search/bing/bing_search.py
@@ -15,15 +15,13 @@ from semantic_kernel.connectors.search.bing.const import (
     QUERY_PARAMETERS,
 )
 from semantic_kernel.connectors.search_engine.bing_connector_settings import BingSettings
-from semantic_kernel.data import (
-    AnyTagsEqualTo,
-    EqualTo,
-    KernelSearchResult,
-    TextSearch,
-    TextSearchFilter,
-    TextSearchOptions,
-    TextSearchResult,
-)
+from semantic_kernel.data.filter_clauses.any_tags_equal_to_filter_clause import AnyTagsEqualTo
+from semantic_kernel.data.filter_clauses.equal_to_filter_clause import EqualTo
+from semantic_kernel.data.kernel_search_result import KernelSearchResult
+from semantic_kernel.data.text_search.text_search import TextSearch
+from semantic_kernel.data.text_search.text_search_filter import TextSearchFilter
+from semantic_kernel.data.text_search.text_search_options import TextSearchOptions
+from semantic_kernel.data.text_search.text_search_result import TextSearchResult
 from semantic_kernel.exceptions import ServiceInitializationError, ServiceInvalidRequestError
 from semantic_kernel.kernel_pydantic import KernelBaseModel
 from semantic_kernel.utils.experimental_decorator import experimental_class
