+++ python/semantic_kernel/data/__init__.py
@@ -1,20 +1,27 @@
 # Copyright (c) Microsoft. All rights reserved.
 
-from semantic_kernel.data.const import DEFAULT_COUNT, DEFAULT_DESCRIPTION
-from semantic_kernel.data.filter_clauses import AnyTagsEqualTo, EqualTo
-from semantic_kernel.data.kernel_search_result import KernelSearchResult
-from semantic_kernel.data.record_definition import (
+from semantic_kernel.data.const import (
+    DEFAULT_COUNT,
+    DEFAULT_DESCRIPTION,
     DistanceFunction,
     IndexKind,
-    VectorStoreRecordDataField,
-    VectorStoreRecordDefinition,
-    VectorStoreRecordKeyField,
-    VectorStoreRecordUtils,
-    VectorStoreRecordVectorField,
-    vectorstoremodel,
 )
+from semantic_kernel.data.filter_clauses.any_tags_equal_to_filter_clause import AnyTagsEqualTo
+from semantic_kernel.data.filter_clauses.equal_to_filter_clause import EqualTo
+from semantic_kernel.data.kernel_search_result import KernelSearchResult
+from semantic_kernel.data.record_definition.vector_store_model_decorator import vectorstoremodel
+from semantic_kernel.data.record_definition.vector_store_model_definition import VectorStoreRecordDefinition
+from semantic_kernel.data.record_definition.vector_store_record_fields import (
+    VectorStoreRecordDataField,
+    VectorStoreRecordKeyField,
+    VectorStoreRecordVectorField,
+)
+from semantic_kernel.data.record_definition.vector_store_record_utils import VectorStoreRecordUtils
 from semantic_kernel.data.search_options_base import SearchOptions
-from semantic_kernel.data.text_search import TextSearch, TextSearchFilter, TextSearchOptions, TextSearchResult
+from semantic_kernel.data.text_search.text_search import TextSearch
+from semantic_kernel.data.text_search.text_search_filter import TextSearchFilter
+from semantic_kernel.data.text_search.text_search_options import TextSearchOptions
+from semantic_kernel.data.text_search.text_search_result import TextSearchResult
 from semantic_kernel.data.vector_search import (
     VectorSearch,
     VectorSearchFilter,
@@ -23,7 +30,8 @@ from semantic_kernel.data.vector_search import (
     VectorSearchQueryTypes,
     VectorSearchResult,
 )
-from semantic_kernel.data.vector_storage import VectorStore, VectorStoreRecordCollection
+from semantic_kernel.data.vector_storage.vector_store import VectorStore
+from semantic_kernel.data.vector_storage.vector_store_record_collection import VectorStoreRecordCollection
 
 __all__ = [
     "DEFAULT_COUNT",
