+++ python/semantic_kernel/data/record_definition/__init__.py
@@ -1,6 +1,5 @@
 # Copyright (c) Microsoft. All rights reserved.
 
-from semantic_kernel.data.record_definition.const import DistanceFunction, IndexKind
 from semantic_kernel.data.record_definition.vector_store_model_decorator import vectorstoremodel
 from semantic_kernel.data.record_definition.vector_store_model_definition import VectorStoreRecordDefinition
 from semantic_kernel.data.record_definition.vector_store_record_fields import (
@@ -11,8 +10,6 @@ from semantic_kernel.data.record_definition.vector_store_record_fields import (
 from semantic_kernel.data.record_definition.vector_store_record_utils import VectorStoreRecordUtils
 
 __all__ = [
-    "DistanceFunction",
-    "IndexKind",
     "VectorStoreRecordDataField",
     "VectorStoreRecordDefinition",
     "VectorStoreRecordKeyField",
