--- python/semantic_kernel/data/record_definition/const.py
@@ -1,20 +1,0 @@
-# Copyright (c) Microsoft. All rights reserved.
-
-
-from enum import Enum
-
-
-class IndexKind(str, Enum):
-    """Index kinds for similarity search."""
-
-    HNSW = "hnsw"
-    FLAT = "flat"
-
-
-class DistanceFunction(str, Enum):
-    """Distance functions for similarity search."""
-
-    COSINE = "cosine"
-    DOT_PROD = "dot_prod"
-    EUCLIDEAN = "euclidean"
-    MANHATTAN = "manhattan"
